refactor entrypoints
